feat(*):remove auction in separate project
