+++ auction/contracts/LibBidDataV1.sol
