+++ auction/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@rarible/auction",
+  "main": "truffle.js",
+  "directories": {
+    "test": "test"
+  },
+  "private": true,
+  "scripts": {
+    "test": "truffle test"
+  },
+  "dependencies": {
+    "@openzeppelin/contracts-upgradeable": "3.4.0",
+    "@rarible/exchange-interfaces": "0.2.0",
+    "@rarible/lib-asset": "0.1.0",
+    "@rarible/royalties": "0.7.2",
+    "@rarible/royalties-upgradeable": "0.7.2"
+  },
+  "devDependencies": {
+    "@daonomic/tests-common": "0.2.2",
+    "@openzeppelin/truffle-upgrades": "^1.8.0",
+    "@truffle/hdwallet-provider": "^1.1.1",
+    "sol-merger": "^3.1.0",
+    "truffle-assertions": "0.9.2",
+    "truffle-plugin-verify": "0.5.7"
+  }
+}
