+++ auction/test-auction.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+truffle test ./test/AuctionHouse.test.js \
+            ./test/contracts/TransferProxyTest.sol \
+            ./test/contracts/ERC20TransferProxyTest.sol \
+            ./test/contracts/TestERC721.sol
+
